Update cortex-m-rt version to prevent ICE
Attempting to build an example would cause an internal compiler
error[0], caused by a change to the `start` language item. Updating to
a newer version of cortex-m-rt fixes the issue.

The dependency to cortex-m-rt was previously pinned to version 0.3.7.
Since the reason for that is not documented, it's not unlikely that I'm
breaking something else that I'm not aware of.

[0]: https://github.com/rust-lang/rust/issues/47040