Add bash scripts for flashing examples
These scripts can be used to upload examples to a board using OpenOCD
and GDB. They work for me, but they're certainly not applicable to all
use cases (for example, on any platform that doesn't have Bash). I
figure it's still nice to have them though, even if to just show a
working configuration that one can apply to their own platform.